switch to complete version off all routers
