+++ src/lsst/cmservice/routers/campaigns.py
@@ -1,7 +1,8 @@
 from collections.abc import Sequence
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, HTTPException
 from safir.dependencies.db_session import db_session_dependency
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import async_scoped_session
 
 from .. import db, models
@@ -25,7 +26,6 @@ router = APIRouter(
 )
 async def get_rows(
     parent_id: int | None = None,
-    parent_name: str | None = None,
     skip: int = 0,
     limit: int = 100,
     session: async_scoped_session = Depends(db_session_dependency),
@@ -35,7 +35,6 @@ async def get_rows(
         parent_id=parent_id,
         skip=skip,
         limit=limit,
-        parent_name=parent_name,
         parent_class=db.Production,
     )
 
@@ -50,3 +49,62 @@ async def get_row(
     session: async_scoped_session = Depends(db_session_dependency),
 ) -> db_class:
     return await db_class.get_row(session, row_id)
+
+
+@router.post(
+    "",
+    status_code=201,
+    response_model=response_model_class,
+    summary=f"Create a {class_string}",
+)
+async def post_row(
+    row_create: create_model_class,
+    session: async_scoped_session = Depends(db_session_dependency),
+) -> db_class:
+    result = await db_class.create_row(session, **row_create.dict())
+    await session.commit()
+    return result
+
+
+@router.delete(
+    "/{row_id}",
+    status_code=204,
+    summary=f"Delete a {class_string}",
+)
+async def delete_row(
+    row_id: int,
+    session: async_scoped_session = Depends(db_session_dependency),
+) -> None:
+    await db_class.delete_row(session, row_id)
+
+
+@router.put(
+    "/{row_id}",
+    response_model=response_model_class,
+    summary=f"Update a {class_string}",
+)
+async def update_row(
+    row_id: int,
+    row_update: response_model_class,
+    session: async_scoped_session = Depends(db_session_dependency),
+) -> db_class:
+    result = await db_class.update_row(session, row_id, **row_update.dict())
+    await session.commit()
+    return result
+
+
+@router.put(
+    "/process/{campaign_id}",
+    response_model=models.Campaign,
+    summary="Process a campaign",
+)
+async def process_campaign(
+    campaign_id: int,
+    session: async_scoped_session = Depends(db_session_dependency),
+) -> db.Campaign:
+    try:
+        campaign = await db.Campaign.get_row(session, campaign_id)
+        await campaign.process(session)
+        return campaign
+    except IntegrityError as e:
+        raise HTTPException(422, detail=str(e)) from e
