+++ src/lsst/cmservice/routers/product_sets.py
